# Faster align()