@krlmlr 2021-01-27
The new implementation seems worse for strings of length > 1.